+++ test/HOneInchV3.test.js
@@ -151,10 +151,12 @@ contract('OneInchV3 Swap', function([_, user]) {
           // sub 1 more percent to tolerate the slippage calculation difference with 1inch
           tokenUser.add(mulPercent(quote, 100 - slippage - 1))
         );
-        expect(await this.token.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(
+          await this.token.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
 
         // Verify ether balance
-        expect(await balanceProxy.get()).to.be.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
         expect(await balanceUser.delta()).to.be.bignumber.eq(
           ether('0')
             .sub(value)
@@ -210,10 +212,12 @@ contract('OneInchV3 Swap', function([_, user]) {
           // sub 1 more percent to tolerate the slippage calculation difference with 1inch
           tokenUser.add(mulPercent(quote, 100 - slippage - 1))
         );
-        expect(await this.token.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(
+          await this.token.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
 
         // Verify ether balance
-        expect(await balanceProxy.get()).to.be.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
         expect(await balanceUser.delta()).to.be.bignumber.eq(
           ether('0')
             .sub(value)
@@ -280,10 +284,12 @@ contract('OneInchV3 Swap', function([_, user]) {
           // sub 1 more percent to tolerate the slippage calculation difference with 1inch
           tokenUser.add(mulPercent(quote, 100 - slippage - 1))
         );
-        expect(await this.token.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(
+          await this.token.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
 
         // Verify ether balance
-        expect(await balanceProxy.get()).to.be.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
         expect(await balanceUser.delta()).to.be.bignumber.eq(
           ether('0')
             .sub(value)
@@ -291,6 +297,49 @@ contract('OneInchV3 Swap', function([_, user]) {
         );
 
         profileGas(receipt);
+      });
+
+      it('should revert: wrong src token amount(ether)', async function() {
+        const value = ether('0.1');
+        const to = this.hOneInch.address;
+        const slippage = 3;
+        const swapReq = queryString.stringifyUrl({
+          url: URL_1INCH_SWAP,
+          query: {
+            fromTokenAddress: '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE',
+            toTokenAddress: tokenAddress,
+            amount: value,
+            slippage: slippage,
+            disableEstimate: true,
+            fromAddress: this.proxy.address,
+            // If the route contains only Uniswap and its' forks, tx.data will invoke `unoswap`
+            protocols: UNOSWAP_PROTOCOLS,
+          },
+        });
+
+        // Call 1inch API
+        const swapResponse = await fetch(swapReq);
+        expect(swapResponse.ok, '1inch api response not ok').to.be.true;
+        const swapData = await swapResponse.json();
+        const quote = swapData.toTokenAmount;
+        // Verify it's `unoswap` function call
+        expect(swapData.tx.data.substring(0, 10)).to.be.eq(
+          SELECTOR_1INCH_UNOSWAP
+        );
+
+        // Prepare handler data
+        const data = getCallData(HOneInch, 'unoswap', [
+          '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE',
+          value.sub(new BN(1000)),
+          tokenAddress,
+          swapData.tx.data,
+        ]);
+
+        // Execute
+        await expectRevert(
+          this.proxy.execMock(to, data, { from: user, value: ether('0.1') }),
+          'HOneInchV3_unoswap: invalid msg.value'
+        );
       });
     });
   });
@@ -367,10 +416,12 @@ contract('OneInchV3 Swap', function([_, user]) {
         expect(await this.token.balanceOf.call(user)).to.be.bignumber.eq(
           tokenUser
         );
-        expect(await this.token.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(
+          await this.token.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
 
         // Verify ether balance
-        expect(await balanceProxy.get()).to.be.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
         expect(balanceUserDelta).to.be.bignumber.gte(
           ether('0')
             // sub 1 more percent to tolerate the slippage calculation difference with 1inch
@@ -445,10 +496,12 @@ contract('OneInchV3 Swap', function([_, user]) {
         expect(await this.token.balanceOf.call(user)).to.be.bignumber.eq(
           tokenUser
         );
-        expect(await this.token.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(
+          await this.token.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
 
         // Verify ether balance
-        expect(await balanceProxy.get()).to.be.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
         expect(balanceUserDelta).to.be.bignumber.gte(
           ether('0')
             // sub 1 more percent to tolerate the slippage calculation difference with 1inch
@@ -541,17 +594,21 @@ contract('OneInchV3 Swap', function([_, user]) {
         expect(await this.token0.balanceOf.call(user)).to.be.bignumber.eq(
           token0User
         );
-        expect(await this.token0.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(
+          await this.token0.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
 
         // Verify token1 balance
         expect(await this.token1.balanceOf.call(user)).to.be.bignumber.gte(
           // sub 1 more percent to tolerate the slippage calculation difference with 1inch
           token1User.add(mulPercent(quote, 100 - slippage - 1))
         );
-        expect(await this.token1.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(
+          await this.token1.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
 
         // Verify ether balance
-        expect(await balanceProxy.get()).to.be.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
         expect(await balanceUser.delta()).to.be.bignumber.eq(
           ether('0').sub(new BN(receipt.receipt.gasUsed))
         );
@@ -621,17 +678,21 @@ contract('OneInchV3 Swap', function([_, user]) {
         expect(await this.token0.balanceOf.call(user)).to.be.bignumber.eq(
           token0User
         );
-        expect(await this.token0.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(
+          await this.token0.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
 
         // Verify token1 balance
         expect(await this.token1.balanceOf.call(user)).to.be.bignumber.gte(
           // sub 1 more percent to tolerate the slippage calculation difference with 1inch
           token1User.add(mulPercent(quote, 100 - slippage - 1))
         );
-        expect(await this.token1.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(
+          await this.token1.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
 
         // Verify ether balance
-        expect(await balanceProxy.get()).to.be.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
         expect(await balanceUser.delta()).to.be.bignumber.eq(
           ether('0').sub(new BN(receipt.receipt.gasUsed))
         );
@@ -699,17 +760,21 @@ contract('OneInchV3 Swap', function([_, user]) {
         expect(await this.token0.balanceOf.call(user)).to.be.bignumber.eq(
           token0User
         );
-        expect(await this.token0.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(
+          await this.token0.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
 
         // Verify weth balance
         expect(await this.weth.balanceOf.call(user)).to.be.bignumber.gte(
           // sub 1 more percent to tolerate the slippage calculation difference with 1inch
           wethUser.add(mulPercent(quote, 100 - slippage - 1))
         );
-        expect(await this.weth.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(
+          await this.weth.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
 
         // Verify ether balance
-        expect(await balanceProxy.get()).to.be.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
         expect(await balanceUser.delta()).to.be.bignumber.eq(
           ether('0').sub(new BN(receipt.receipt.gasUsed))
         );
@@ -760,7 +825,7 @@ contract('OneInchV3 Swap', function([_, user]) {
           WETH_TOKEN,
           swapData.tx.data,
         ]);
-        var data = data + appendData;
+        data = data + appendData;
 
         // Execute
         const receipt = await this.proxy.execMock(to, data, {
@@ -779,17 +844,21 @@ contract('OneInchV3 Swap', function([_, user]) {
         expect(await this.token0.balanceOf.call(user)).to.be.bignumber.eq(
           token0User
         );
-        expect(await this.token0.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(
+          await this.token0.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
 
         // Verify weth balance
         expect(await this.weth.balanceOf.call(user)).to.be.bignumber.gte(
           // sub 1 more percent to tolerate the slippage calculation difference with 1inch
           wethUser.add(mulPercent(quote, 100 - slippage - 1))
         );
-        expect(await this.weth.balanceOf.call(this.proxy.address)).to.be.zero;
+        expect(
+          await this.weth.balanceOf.call(this.proxy.address)
+        ).to.be.bignumber.zero;
 
         // Verify ether balance
-        expect(await balanceProxy.get()).to.be.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
         expect(await balanceUser.delta()).to.be.bignumber.eq(
           ether('0').sub(new BN(receipt.receipt.gasUsed))
         );
@@ -892,6 +961,152 @@ contract('OneInchV3 Swap', function([_, user]) {
         await expectRevert(
           this.proxy.execMock(to, data, { from: user, value: ether('0.1') }),
           'HOneInchV3_unoswap: Invalid output token amount'
+        );
+      });
+
+      it('should revert: wrong src token(ether)', async function() {
+        // Prepare data
+        const value = ether('100');
+        const to = this.hOneInch.address;
+        const slippage = 3;
+        const swapReq = queryString.stringifyUrl({
+          url: URL_1INCH_SWAP,
+          query: {
+            fromTokenAddress: token0Address,
+            toTokenAddress: token1Address,
+            amount: value,
+            slippage: slippage,
+            disableEstimate: true,
+            fromAddress: this.proxy.address,
+            // If the route contains only Uniswap and its' forks, tx.data will invoke `unoswap`
+            protocols: UNOSWAP_PROTOCOLS,
+          },
+        });
+
+        // Transfer from token to Proxy first
+        await this.token0.transfer(this.proxy.address, value, {
+          from: providerAddress,
+        });
+        await this.proxy.updateTokenMock(this.token0.address);
+
+        // Call 1inch API
+        const swapResponse = await fetch(swapReq);
+        expect(swapResponse.ok, '1inch api response not ok').to.be.true;
+        const swapData = await swapResponse.json();
+        const quote = swapData.toTokenAmount;
+        // Verify it's `unoswap` function call
+        expect(swapData.tx.data.substring(0, 10)).to.be.eq(
+          SELECTOR_1INCH_UNOSWAP
+        );
+
+        // Prepare handler data
+        const data = getCallData(HOneInch, 'unoswap', [
+          '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE',
+          value,
+          token1Address,
+          swapData.tx.data,
+        ]);
+
+        await expectRevert.unspecified(
+          this.proxy.execMock(to, data, { from: user, value: ether('0.1') })
+        );
+      });
+
+      it('should revert: wrong src token(erc20)', async function() {
+        // Prepare data
+        const value = ether('100');
+        const to = this.hOneInch.address;
+        const slippage = 3;
+        const swapReq = queryString.stringifyUrl({
+          url: URL_1INCH_SWAP,
+          query: {
+            fromTokenAddress: token0Address,
+            toTokenAddress: token1Address,
+            amount: value,
+            slippage: slippage,
+            disableEstimate: true,
+            fromAddress: this.proxy.address,
+            // If the route contains only Uniswap and its' forks, tx.data will invoke `unoswap`
+            protocols: UNOSWAP_PROTOCOLS,
+          },
+        });
+
+        // Transfer from token to Proxy first
+        await this.token0.transfer(this.proxy.address, value, {
+          from: providerAddress,
+        });
+        await this.proxy.updateTokenMock(this.token0.address);
+
+        // Call 1inch API
+        const swapResponse = await fetch(swapReq);
+        expect(swapResponse.ok, '1inch api response not ok').to.be.true;
+        const swapData = await swapResponse.json();
+        const quote = swapData.toTokenAmount;
+        // Verify it's `unoswap` function call
+        expect(swapData.tx.data.substring(0, 10)).to.be.eq(
+          SELECTOR_1INCH_UNOSWAP
+        );
+
+        // Prepare handler data
+        const data = getCallData(HOneInch, 'unoswap', [
+          WETH_TOKEN,
+          value,
+          token1Address,
+          swapData.tx.data,
+        ]);
+
+        await expectRevert(
+          this.proxy.execMock(to, data, { from: user, value: ether('0.1') }),
+          'HOneInchV3_unoswap: Dai/insufficient-allowance'
+        );
+      });
+
+      it('should revert: wrong src token amount(erc20)', async function() {
+        // Prepare data
+        const value = ether('100');
+        const to = this.hOneInch.address;
+        const slippage = 3;
+        const swapReq = queryString.stringifyUrl({
+          url: URL_1INCH_SWAP,
+          query: {
+            fromTokenAddress: token0Address,
+            toTokenAddress: token1Address,
+            amount: value,
+            slippage: slippage,
+            disableEstimate: true,
+            fromAddress: this.proxy.address,
+            // If the route contains only Uniswap and its' forks, tx.data will invoke `unoswap`
+            protocols: UNOSWAP_PROTOCOLS,
+          },
+        });
+
+        // Transfer from token to Proxy first
+        await this.token0.transfer(this.proxy.address, value, {
+          from: providerAddress,
+        });
+        await this.proxy.updateTokenMock(this.token0.address);
+
+        // Call 1inch API
+        const swapResponse = await fetch(swapReq);
+        expect(swapResponse.ok, '1inch api response not ok').to.be.true;
+        const swapData = await swapResponse.json();
+        const quote = swapData.toTokenAmount;
+        // Verify it's `unoswap` function call
+        expect(swapData.tx.data.substring(0, 10)).to.be.eq(
+          SELECTOR_1INCH_UNOSWAP
+        );
+
+        // Prepare handler data
+        const data = getCallData(HOneInch, 'unoswap', [
+          token0Address,
+          value.sub(new BN(1000)),
+          token1Address,
+          swapData.tx.data,
+        ]);
+
+        await expectRevert(
+          this.proxy.execMock(to, data, { from: user, value: ether('0.1') }),
+          'HOneInchV3_unoswap: Dai/insufficient-allowance'
         );
       });
     });