+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 
+- Fix 1inch v3 unoswap wrong dst token issue.
+
 ## [1.8.0] - 2021-09-13
 
 ### Added